+++ src/test/java/domain/StatisticsServiceTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import service.StatisticsService;
 
 class StatisticsServiceTest {
     private final StatisticsService statisticsService = new StatisticsService();
